+++ R/SAM2OM.R
@@ -38,26 +38,30 @@ SAM2OM<-function(fit, nsim=32, proyears=30, interval=1, Name = "SAM Model", WLa=
     stop("Install the TMB and mvtnorm packages to use SAM2OM")
   }
 
+  if(!requireNamespace("stockassessment", quietly = TRUE)) {
+    stop("Install the `stockassessment` package to use SAM2OM. \n`pak::pkg_install('fishfollower/SAM/stockassessment')`")
+  }
+  
   vanilla_SAM = is.null(fit$input) # is this SAM or some earlier version / modified version (e.g. Chub Mackerel)
   
   if(vanilla_SAM){
     
-    dims = dim(faytable(fit))
+    dims = dim(stockassessment::faytable(fit))
     ny = dims[1]
     na = dims[2]
-    yrs = row.names(faytable(fit))
+    yrs = row.names(stockassessment::faytable(fit))
     CurrentYr = max(as.numeric(yrs))
     waa = array(rep(t(fit$data$stockMeanWeight),each=nsim),c(nsim,na,ny)) # fit$input$dat$waa #weight at age
     Mataa = array(rep(t(fit$data$propMat),each=nsim),c(nsim,na,ny)) # fit$input$dat$maa #maturity at age
     Maa = array(rep(t(fit$data$natMor),each=nsim),c(nsim,na,ny)) # fit$input$dat$M #M at age
     laa = (waa/WLa)^(1/WLb) # length at age isn't used unless catch at length and mean length are simulated
-    caa = array(rep(t(caytable(fit)),each=nsim),c(nsim,na,ny)) # fit$caa #estimated catch at age
+    caa = array(rep(t(stockassessment::caytable(fit)),each=nsim),c(nsim,na,ny)) # fit$caa #estimated catch at age
     
     
     if(!stoch){ # deterministic historical reconstruction
       
-      faa = array(rep(t(faytable(fit)),each=nsim),c(nsim,na,ny)) # f at age
-      naa = array(rep(t(ntable(fit)),each=nsim),c(nsim,na,ny)) # number at age (million)
+      faa = array(rep(t(stockassessment::faytable(fit)),each=nsim),c(nsim,na,ny)) # f at age
+      naa = array(rep(t(stockassessment::ntable(fit)),each=nsim),c(nsim,na,ny)) # number at age (million)
         
     }else{ # rerun SAM 
       
@@ -77,13 +81,13 @@ SAM2OM<-function(fit, nsim=32, proyears=30, interval=1, Name = "SAM Model", WLa=
       ind_logN <- SD$jointPrecision@Dimnames[[1]] == "logN"
       cov_logN <- SD$jointPrecision[ind_logN, ind_logN] %>% as.matrix()
      
-      log_NAA <- rmvnorm(nsim, as.numeric(mean_logN), cov_logN) %>%
-        array(c(nsim, n_age, nscodData$noYears))
+      log_NAA <- mvtnorm::rmvnorm(nsim, as.numeric(mean_logN), cov_logN) %>%
+        array(c(nsim, n_age, stockassessment::nscodData$noYears))
       
-      MAA <- nscodData$natMor
+      MAA <- stockassessment::nscodData$natMor
       FAA <- array(NA, dim(log_NAA))
       
-      for (y in 2:nscodData$noYears - 1) {
+      for (y in 2:stockassessment::nscodData$noYears - 1) {
         SAA <- log_NAA[, 2:n_age, y+1]-log_NAA[, 2:n_age - 1, y] # survival
         ZAA <- -log(SAA)                                           # Z
         FAA[, 2:n_age - 1, y] <- ZAA - matrix(MAA[y, 2:n_age - 1], nsim, n_age - 1, byrow = TRUE)